fix: player health slider

+++ Assets/Scripts/Models/LocalData/MainLocalData.cs
@@ -1,12 +1,11 @@
 namespace Models.LocalData
 {
-    using System.Numerics;
     using GDK.LocalData;
 
     public class MainLocalData : ILocalData
     {
         public string      Username    { get; set; }
-        public BigInteger  Coins       { get; set; }
+        public long        Coins       { get; set; }
         public UpgradeData UpgradeData { get; set; }
 
         public void Initialize()
