+++ Assets/Scripts/Signals/UpgradedStatSignal.cs
@@ -5,6 +5,6 @@ namespace Signals
     public class UpgradedStatSignal
     {
         public StatType StatType { get; init; }
-        public uint     Level    { get; init; }
+        public long     Level    { get; init; }
     }
 }
